# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: LeadForge AI REST API
  version: 1.0.0
  description: |
    Complete REST API for LeadForge AI.
    All endpoints are prefixed with `/api/v1`.
    Two auth modes:
    1. **Clerk JWT** — short-lived Bearer token from the Clerk session. Optional `X-Team-Id` header selects active team; defaults to first membership.
    2. **API Key** — long-lived `lf_*` token (Bearer). Bound to a single team. Listed/managed at `/api-keys`.
servers:
  - url: https://api.leadforge.ai/api/v1
  - url: http://localhost:4000/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or lf_api_key
  parameters:
    teamHeader:
      name: X-Team-Id
      in: header
      schema: { type: string }
      description: Active team (Clerk JWT only).
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [UNAUTHORIZED, BAD_REQUEST, RATE_LIMITED, PAYMENT_REQUIRED, NOT_FOUND] }
            message: { type: string }
            details: { type: object }
...
    JobStatus: { type: string, enum: [PENDING, QUEUED, RUNNING, PAUSED, COMPLETED, FAILED, CANCELLED] }
    SourceType:
      type: string
      enum: [WEB_SEARCH, DIRECTORY, COMPANY_PAGE, BLOG, FORUM, SOCIAL_LINKEDIN, SOCIAL_TWITTER, CONTACT_PAGE, LISTING, DATABASE, CUSTOM_URL_LIST]
    VerificationStatus: { type: string, enum: [PENDING, VALID, INVALID, RISKY, UNKNOWN, CATCH_ALL] }
    ExportFormat: { type: string, enum: [CSV, XLSX, JSON] }
    FilterCondition:
      type: object
      required: [field, operator, value]
      properties:
        field:
          type: string
          enum: [niche, industry, domain_extension, country, region, city, language, keyword, companyName, companyDomain, companySize, companyRevenue, companyAuthority, jobTitle, jobRole, jobSeniority, jobDepartment, technology, social_presence, has_email, has_phone, has_contact_page, qualityScore, relevanceScore, intentScore, authorityScore, verificationStatus, createdAt, sourceType]
        operator:
          type: string
          enum: [eq, neq, in, nin, contains, starts_with, ends_with, gt, gte, lt, lte, between, exists, not_exists, has_any, has_all]
        value: {}
    Filter:
      oneOf:
        - $ref: '#/components/schemas/FilterCondition'
        - type: object
          properties:
            AND: { type: array, items: { $ref: '#/components/schemas/Filter' } }
        - type: object
          properties:
            OR: { type: array, items: { $ref: '#/components/schemas/Filter' } }
        - type: object
          properties:
            NOT: { $ref: '#/components/schemas/Filter' }
...
    Job:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        progress: { type: number }
        leadsFound: { type: integer }
        leadsVerified: { type: integer }
        targetLeads: { type: integer }
        sources: { type: array, items: { $ref: '#/components/schemas/SourceType' } }
        filters: { $ref: '#/components/schemas/Filter' }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
...
paths:
...
  /jobs:
    post:
      summary: Create extraction job
      parameters: [{ $ref: '#/components/parameters/teamHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, sources, filters]
              properties:
                name: { type: string }
                description: { type: string }
                sources: { type: array, items: { $ref: '#/components/schemas/SourceType' } }
                filters: { $ref: '#/components/schemas/Filter' }
                targetLeads: { type: integer, default: 100, maximum: 50000 }
                priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
      responses:
        '201':
          description: created
          content: { application/json: { schema: { type: object, properties: { job: { $ref: '#/components/schemas/Job' } } } } }
        '402': { description: insufficient credits, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    get:
      summary: List jobs
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/JobStatus' }
      responses: { '200': { description: ok } }
  /jobs/{id}:
    get: { summary: Get job, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /jobs/{id}/pause:
    post: { summary: Pause job, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /jobs/{id}/resume:
    post: { summary: Resume job, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /jobs/{id}/cancel:
    post: { summary: Cancel job, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
...
  /leads:
    get:
      summary: List leads
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: jobId, in: query, schema: { type: string } }
        - { name: verificationStatus, in: query, schema: { $ref: '#/components/schemas/VerificationStatus' } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, qualityScore, companyName, fullName] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses: { '200': { description: ok } }
...
  /leads/{id}:
    get: { summary: Get lead, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
    patch:
      summary: Update lead
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [NEW, VERIFIED, ENRICHED, EXPORTED, ARCHIVED] }
                isFavorite: { type: boolean }
                notes: { type: string, nullable: true }
      responses: { '200': { description: ok } }
    delete: { summary: Delete lead, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '204': { description: deleted } } }
...
  /exports:
    post:
      summary: Create export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format]
              properties:
                format: { $ref: '#/components/schemas/ExportFormat' }
                filter: { $ref: '#/components/schemas/Filter' }
                leadIds: { type: array, items: { type: string } }
                jobId: { type: string }
      responses: { '202': { description: queued } }
    get: { summary: List exports, responses: { '200': { description: ok } } }
  /exports/{id}:
    get: { summary: Get export status, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
...
  /api-keys:
    post:
      summary: Create API key (returned once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                scopes: { type: array, items: { type: string } }
                expiresAt: { type: string, format: date-time }
      responses: { '201': { description: created } }
    get: { summary: List API keys, responses: { '200': { description: ok } } }
  /api-keys/{id}:
    delete: { summary: Revoke API key, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '204': { description: revoked } } }
...